Replace http::uri::Uri with url::Url
The version of the `http::uri::Uri` that is used in this project is old,
because it is provided by Actix. It was only directly used by the
project to generate enrollment URLs. This also required the `http-serde`
library at an older version to be compatible. This is not ideal, so I
replaced it with the latest version of the `url::Url` which does the job
even better.