Improve error messages reported by the program
At the moment, we only print the first (outermost) error message when
the program fails. This is not very helpful and results in messages
like:
```
2025-10-15T09:18:18.631953Z ERROR main mctrlrs: 103: Failed to load
configuration file
```

as this tells what happened, but not why. With this change we will print
the entire error stack preserving the context in full:

```
2025-10-15T09:20:47.187633Z ERROR main mctrlrs: 101: Failed to load
configuration file: Failed to parse configuration file: missing field
`server_properties_path` at line 2 column 1
```