Add support for stable cookie signing key
Prior to this change, every time the server is restarted, it would
generate a new cookie signing key. This means that any active session
will be immediately invalidated.

With this change, we now can configure a stable master key that will be
reused. This is still insufficient to maintain the session after the
server restart, because we lose session state, as it is only stored
in-memory. Persisent sessions will be implemented next.